Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
  int main () {
-    int a,b,c;
-    int ashi=(a+b+c)/3;
-    scanf("%d",&a);
-    scanf("%d",&b);
-    scanf("%d",&c);
+    float a,b,c;
+    float ashi=(a+b+c)/3;
+    scanf("%f",&a);
+    scanf("%f",&b);
+    scanf("%f",&c);
     printf("Average: %.2f",ashi);
